# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entitlements/chosen-language:
    post:
      operationId: setChosenLanguage
      tags: [entitlements]
      summary: Record the One Language tier's chosen language
      description: >-
        Records the single language a One Language ($6.99) subscriber unlocks on
        top of free Hindi, captured at purchase. Once set while on the middle
        tier the choice is locked: changing it returns 409, and only upgrading
        to all-access frees it. Choosing Hindi or an unknown language is
        rejected. Returns the fresh entitlements snapshot.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetChosenLanguageInput"
      responses:
        "200":
          description: The caller's updated entitlements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entitlements"
        "400":
          description: Missing or invalid language
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Language not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The chosen language is locked for this subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SetChosenLanguageInput:
      type: object
      properties:
        language:
          type: string
          description: The language code to unlock (must not be Hindi).
      required:
        - language
...
    PlanFeatures:
      type: object
      properties:
        allLanguages:
          type: boolean
        unlimitedLessons:
          type: boolean
        review:
          type: boolean
        advancedAnalytics:
          type: boolean
        extendedLibrary:
          type: boolean
          description: >-
            Whether the caller can access the deep, pre-seeded premium phrase
            library beyond each topic's free starter set. Plus only.
        unlimitedChatTime:
          type: boolean
          description: >-
            Whether the caller has no weekly ceiling on Bolo Parrot
            conversational chat time. Free is capped at 2 minutes/week; One
            Language and Plus are unlimited.
      required:
        - allLanguages
        - unlimitedLessons
        - review
        - advancedAnalytics
        - extendedLibrary
        - unlimitedChatTime
    DailyLessonAllowance:
      type: object
      properties:
        limit:
          type: integer
          nullable: true
          description: Daily new-lesson ceiling; null means unlimited (Plus).
        used:
          type: integer
        remaining:
          type: integer
          nullable: true
          description: Remaining today; null means unlimited (Plus).
      required:
        - limit
        - used
        - remaining
    WeeklyChatAllowance:
      type: object
      properties:
        limit:
          type: integer
          nullable: true
          description: >-
            Weekly Bolo Parrot chat-time ceiling in seconds; null means
            unlimited (One Language and Plus).
        used:
          type: integer
        remaining:
          type: integer
          nullable: true
          description: Remaining this week; null means unlimited.
      required:
        - limit
        - used
        - remaining
    EntitlementLimits:
      type: object
      properties:
        dailyNewLessons:
          $ref: "#/components/schemas/DailyLessonAllowance"
        weeklyChatSeconds:
          $ref: "#/components/schemas/WeeklyChatAllowance"
      required:
        - dailyNewLessons
        - weeklyChatSeconds
    Entitlements:
      type: object
      properties:
        plan:
          type: string
          description: The effective plan ("free", "one_language", or "plus").
        status:
          type: string
          description: none | trialing | active | expired | canceled
        trialEndsAt:
          type: string
          format: date-time
          nullable: true
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
        chosenLanguage:
          type: string
          nullable: true
          description: >-
            The single language a One Language ($6.99) subscriber unlocked on
            top of free Hindi. Null for Free and all-access Plus.
        allowedLanguages:
          type: array
          items:
            type: string
          description: The concrete language codes the caller may access.
        features:
          $ref: "#/components/schemas/PlanFeatures"
        limits:
          $ref: "#/components/schemas/EntitlementLimits"
      required:
        - plan
        - status
        - trialEndsAt
        - currentPeriodEnd
        - chosenLanguage
        - allowedLanguages
        - features
        - limits
...